pyproject.toml
[project]
name = "osm-lpv"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "argparse>=1.4.0",
    "python-crontab>=3.2.0",
]
